Heavy optimizations

+++ Engine/ECS/Systems/EntityStripeSet.cs
@@ -9,12 +9,21 @@ namespace Engine.ECS.Systems;
 public sealed class EntityStripeSet
 {
     private readonly List<int>[] _buckets;
-    private readonly Dictionary<int, (byte Stripe, int IndexInBucket)> _locationsByEntityId = [];
-    private readonly byte _stripeCount;
+    private readonly Dictionary<int, (ushort Stripe, int IndexInBucket)> _locationsByEntityId = [];
+    private readonly ushort _stripeCount;
 
-    public byte StripeCount => _stripeCount;
+    /// <summary>
+    /// ushort, not byte, because TieredEntityStripeSet multiplies a system's own base stripe
+    /// count by its tier's divisor to build each tier's bucket -- 60 (GameTiming.FramesPerSecond,
+    /// used by the regen/body-part systems) times a divisor of 8 is 480, which silently truncated
+    /// to 224 while this was a byte. That made a Beyond-tier entity visited every 224 frames while
+    /// systems computing framesPerVisit in int arithmetic still scaled their effect by 480 -- see
+    /// SimpleHealthRegenSystem. byte * byte tops out at 65,025, so ushort covers every divisor a
+    /// byte-valued ISystem.StripeCount can produce.
+    /// </summary>
+    public ushort StripeCount => _stripeCount;
 
-    public EntityStripeSet(byte stripeCount, ReadOnlySpan<int> existingEntityIds)
+    public EntityStripeSet(ushort stripeCount, ReadOnlySpan<int> existingEntityIds)
     {
         ArgumentOutOfRangeException.ThrowIfZero(stripeCount);
 
@@ -34,7 +43,7 @@ public sealed class EntityStripeSet
     /// <summary>Builds an EntityStripeSet already wired to drivingPool's EntityAdded/EntityRemoved membership events -- the construct-then-subscribe dance every non-tiered EntityStripeSet consumer's constructor would otherwise repeat by hand.</summary>
     /// <param name="stripeCount">The number of stripes to divide entities across.</param>
     /// <param name="drivingPool">The pool whose membership the stripe set should track.</param>
-    public static EntityStripeSet CreateAndWire(byte stripeCount, IEntityMembershipPool drivingPool)
+    public static EntityStripeSet CreateAndWire(ushort stripeCount, IEntityMembershipPool drivingPool)
     {
         var stripeSet = new EntityStripeSet(stripeCount, drivingPool.EntityIds);
         drivingPool.EntityAdded += stripeSet.OnEntityAdded;
@@ -45,13 +54,13 @@ public sealed class EntityStripeSet
     /// <summary>Gets the entities assigned to the given stripe.</summary>
     /// <param name="stripeIndex">The index of the stripe to retrieve entities for.</param>
     /// <returns>A read-only span containing the entities in the specified stripe.</returns>
-    public ReadOnlySpan<int> GetBucket(byte stripeIndex) => CollectionsMarshal.AsSpan(_buckets[stripeIndex]);
+    public ReadOnlySpan<int> GetBucket(ushort stripeIndex) => CollectionsMarshal.AsSpan(_buckets[stripeIndex]);
 
     /// <summary>Handles the addition of a new entity to the set.</summary>
     /// <param name="entityId">The ID of the entity being added.</param>
     public void OnEntityAdded(int entityId)
     {
-        var stripe = (byte)(entityId % _stripeCount);
+        var stripe = (ushort)(entityId % _stripeCount);
         var bucket = _buckets[stripe];
 
         _locationsByEntityId[entityId] = (stripe, bucket.Count);
